requirements.txt
requests==2.32.3
beautifulsoup4==4.13.4
biopython==1.85
PyMuPDF==1.24.5
torch>=2.0.0
ultralytics>=8.0.0
opencv-python>=4.5.0
pillow>=8.0.0
numpy>=1.20.0